Move files and remove unused `UniqueTagList`
