--- src/main/java/seedu/commando/model/tag/UniqueTagList.java
@@ -1,143 +1,0 @@
-package seedu.commando.model.tag;
-
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import seedu.commando.commons.exceptions.IllegalValueException;
-import seedu.commando.commons.util.CollectionUtil;
-
-import java.util.*;
-
-/**
- * A list of tags that enforces no nulls and uniqueness between its elements.
- *
- * Supports minimal set of list operations for the app's features.
- *
- * @see Tag#equals(Object)
- * @see CollectionUtil#elementsAreUnique(Collection)
- */
-public class UniqueTagList implements Iterable<Tag> {
-
-    /**
-     * Signals that an operation would have violated the 'no duplicates' property of the list.
-     */
-    public static class DuplicateTagException extends IllegalValueException {
-        protected DuplicateTagException() {
-            super("Operation would result in duplicate tags");
-        }
-    }
-
-    private final ObservableList<Tag> internalList = FXCollections.observableArrayList();
-
-    /**
-     * Constructs empty TagList.
-     */
-    public UniqueTagList() {}
-
-    /**
-     * Varargs/array constructor, enforces no nulls or duplicates.
-     */
-    public UniqueTagList(Tag... tags) throws DuplicateTagException {
-        assert !CollectionUtil.isAnyNull((Object[]) tags);
-        final List<Tag> initialTags = Arrays.asList(tags);
-        if (!CollectionUtil.elementsAreUnique(initialTags)) {
-            throw new DuplicateTagException();
-        }
-        internalList.addAll(initialTags);
-    }
-
-    /**
-     * java collections constructor, enforces no null or duplicate elements.
-     */
-    public UniqueTagList(Collection<Tag> tags) throws DuplicateTagException {
-        CollectionUtil.assertNoNullElements(tags);
-        if (!CollectionUtil.elementsAreUnique(tags)) {
-            throw new DuplicateTagException();
-        }
-        internalList.addAll(tags);
-    }
-
-    /**
-     * java set constructor, enforces no nulls.
-     */
-    public UniqueTagList(Set<Tag> tags) {
-        CollectionUtil.assertNoNullElements(tags);
-        internalList.addAll(tags);
-    }
-
-    /**
-     * Copy constructor, insulates from changes in source.
-     */
-    public UniqueTagList(UniqueTagList source) {
-        internalList.addAll(source.internalList); // insulate internal list from changes in argument
-    }
-
-    /**
-     * All tags in this list as a Set. This set is mutable and change-insulated against the internal list.
-     */
-    public Set<Tag> toSet() {
-        return new HashSet<>(internalList);
-    }
-
-    /**
-     * Replaces the Tags in this list with those in the argument tag list.
-     */
-    public void setTags(UniqueTagList replacement) {
-        this.internalList.clear();
-        this.internalList.addAll(replacement.internalList);
-    }
-
-    /**
-     * Adds every tag from the argument list that does not yet exist in this list.
-     */
-    public void mergeFrom(UniqueTagList tags) {
-        final Set<Tag> alreadyInside = this.toSet();
-        for (Tag tag : tags) {
-            if (!alreadyInside.contains(tag)) {
-                internalList.add(tag);
-            }
-        }
-    }
-
-    /**
-     * Returns true if the list contains an equivalent Tag as the given argument.
-     */
-    public boolean contains(Tag toCheck) {
-        assert toCheck != null;
-        return internalList.contains(toCheck);
-    }
-
-    /**
-     * Adds a Tag to the list.
-     *
-     * @throws DuplicateTagException if the Tag to add is a duplicate of an existing Tag in the list.
-     */
-    public void add(Tag toAdd) throws DuplicateTagException {
-        assert toAdd != null;
-        if (contains(toAdd)) {
-            throw new DuplicateTagException();
-        }
-        internalList.add(toAdd);
-    }
-
-    @Override
-    public Iterator<Tag> iterator() {
-        return internalList.iterator();
-    }
-
-    public ObservableList<Tag> getInternalList() {
-        return internalList;
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        return other == this // short circuit if same object
-                || (other instanceof UniqueTagList // instanceof handles nulls
-                && this.internalList.equals(
-                ((UniqueTagList) other).internalList));
-    }
-
-    @Override
-    public int hashCode() {
-        return internalList.hashCode();
-    }
-}
